+++ src/main/java/com/academia/bookshop/controller/BookController.java
@@ -66,7 +66,7 @@ public class BookController {
     @PreAuthorize("hasAuthority('" + AUTHORITY_ALL + "') or hasAuthority('" + AUTHORITY_DELETE + "')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PostMapping(value = "/remove", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public void remove(@RequestBody @Parameter(description = "List of book ids") List<Long> ids) {
+    public void remove(@RequestBody @Parameter(description = "List of book ids") List<Integer> ids) {
         bookService.remove(ids);
     }
 }
