+++ src/main/java/com/academia/bookshop/controller/TagController.java
@@ -58,7 +58,7 @@ public class TagController {
     @PreAuthorize("hasAuthority('" + AUTHORITY_ALL + "') or hasAuthority('" + AUTHORITY_DELETE + "')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PostMapping(value = "/remove", consumes = MediaType.APPLICATION_JSON_VALUE)
-    public void remove(@RequestBody @Parameter(description = "List of tag ids") List<Long> ids) {
+    public void remove(@RequestBody @Parameter(description = "List of tag ids") List<Short> ids) {
         tagService.remove(ids);
     }
 }
